deleting update

+++ kanban.server/server/services/ListsService.js
@@ -30,8 +30,8 @@ class ListsService {
     return data
   }
 
-  async deleteList(id) {
-    const data = await dbContext.List.findOneAndDelete({ _id: id })
+  async deleteList(id, userId) {
+    const data = await dbContext.List.findOneAndDelete({ _id: id, creatorId: userId })
     if (!data) {
       throw new BadRequest('Invalid Id')
     }
